Create a connection var in the test
This makes it much easier to evaluate code in Vim over using a fixture,
and it accomplishes the exact same thing, as far as I can tell. So, for
instance, I can put my cursor inside an `is` form, type `cpp`, and get
the result of that test immediately. With fixtures the `connection` var
is not bound to an actual `connection` object, so it fails.