Have ping-pong/stop actually stop the process
Before we were closing the channel returned by go, which, as far as I
can tell, only *closes the channel* (which makes total sense). So now we
will use a separate channel which we will put to when we need the
process to terminate.

Now we don't even need a reference to the channel returned by go because
there is nothing that we can do with it.