Make parse pure, use hash instead of :uuid
Message path will now be traced by its hash, so that parsing a message
multiple times will always return the same map.

This has potential for confusion should an IRC user type the exact same
content in messages that are close together, but I think that would
still be easy enough to figure out, and it's very unlikely to occur
during debugging.